+++ CHANGELOG.md
@@ -1,4 +1,10 @@
 # CHANGELOG
+
+## 0.0.51
+- [common] Override MinecraftVersion#toString
+- [common] Add MinecraftVersion#getByProtocolVersion
+- [common] Add 26.1.x to MinecraftVersion
+
 ## 0.0.50
 - [commands] Improve ArgumentMapper injection
 - [commands] Add HelpMenu